Extracted a Downloader interface and now using a more proper Decorator pattern
